Move `decodeFromResources` to common package
